mcp-okta-backend/pyproject.toml
[project]
name = "mcp-okta-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "pyjwt[crypto]>=2.10.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
